Conflicts:
	codec/encoder/core/inc/rc.h
	codec/encoder/core/src/ratectl.cpp

+++ codec/encoder/core/inc/rc.h
@@ -175,8 +175,8 @@ int32_t   iBitRate;
 int32_t   iPreviousBitrate;
 int32_t   iPreviousGopSize;
 double    fFrameRate;
-int64_t   iBitsPerFrame;
-int64_t   iMaxBitsPerFrame;
+int32_t   iBitsPerFrame;
+int32_t   iMaxBitsPerFrame;
 double    dPreviousFps;
 
 // bits allocation and status
@@ -227,7 +227,7 @@ int32_t   iLastCalculatedQScale;
 //for skip frame and padding
 int32_t   iBufferSizeSkip;
 int64_t   iBufferFullnessSkip;
-int32_t   iBufferMaxBRFullness[TIME_WINDOW_TOTAL];//0: EVEN_TIME_WINDOW; 1: ODD_TIME_WINDOW
+int64_t   iBufferMaxBRFullness[TIME_WINDOW_TOTAL];//0: EVEN_TIME_WINDOW; 1: ODD_TIME_WINDOW
 int32_t   iPredFrameBit;
 bool      bNeedShiftWindowCheck[TIME_WINDOW_TOTAL];
 int32_t   iBufferSizePadding;
